triggerGetInfo вызывается в historyPage
